+++ lib/trellis/utils/output.py
@@ -8,7 +8,7 @@ import re
 import textwrap
 
 from ansible import __version__
-from ansible.utils.unicode import to_unicode
+from ansible.module_utils._text import to_text
 
 def system(vagrant_version=None):
     # Get most recent Trellis CHANGELOG entry
@@ -27,9 +27,9 @@ def system(vagrant_version=None):
 
         # Retrieve most recent changelog entry
         else:
-            change = re.search(r'.*\n\*\s*([^\(\n\[]+)', str)
+            change = re.search(r'^\*\s?(\[BREAKING\])?([^\(\n\[]+)', str, re.M|re.I)
             if change is not None:
-                changelog_msg = '\n  Trellis at "{0}"'.format(change.group(1).strip())
+                changelog_msg = '\n  Trellis at "{0}"'.format(change.group(2).strip())
 
     # Vagrant info, if available
     vagrant = ' Vagrant {0};'.format(vagrant_version) if vagrant_version else ''
@@ -42,15 +42,19 @@ def reset_task_info(obj, task=None):
     obj.first_host = True
     obj.first_item = True
     obj.task_failed = False
-    obj.vagrant_version = None
 
 # Display dict key only, instead of full json dump
 def replace_item_with_key(obj, result):
-    if not obj._display.verbosity:
-        if 'key' in result._result['item']:
-            result._result['item'] = result._result['item']['key']
-        elif 'item' in result._result['item'] and 'key' in result._result['item']['item']:
-            result._result['item'] = result._result['item']['item']['key']
+    if not obj._display.verbosity and 'label' not in result._task._ds.get('loop_control', {}):
+        item = '_ansible_item_label' if '_ansible_item_label' in result._result else 'item'
+        if 'key' in result._result[item]:
+            result._result[item] = result._result[item]['key']
+        elif type(result._result[item]) is dict:
+            subitem = '_ansible_item_label' if '_ansible_item_label' in result._result[item] else 'item'
+            if 'key' in result._result[item].get(subitem, {}):
+                result._result[item] = result._result[item][subitem]['key']
+            elif '_ansible_item_label' in result._result[item]:
+                result._result[item] = result._result[item]['_ansible_item_label']
 
 def display(obj, result):
     msg = ''
@@ -58,10 +62,9 @@ def display(obj, result):
     display = obj._display.display
     wrap_width = 77
     first = obj.first_host and obj.first_item
-    failed = 'failed' in result or 'unreachable' in result
 
     # Only display msg if debug module or if failed (some modules have undesired 'msg' on 'ok')
-    if 'msg' in result and (failed or obj.action == 'debug'):
+    if 'msg' in result and (obj.task_failed or obj.action == 'debug'):
         msg = result.pop('msg', '')
 
         # Disable Ansible's verbose setting for debug module to avoid the CallbackBase._dump_results()
@@ -69,8 +72,8 @@ def display(obj, result):
             del result['_ansible_verbose_always']
 
     # Display additional info when failed
-    if failed:
-        items = (item for item in ['reason', 'module_stderr', 'module_stdout', 'stderr'] if item in result and to_unicode(result[item]) != '')
+    if obj.task_failed:
+        items = (item for item in ['reason', 'module_stderr', 'module_stdout', 'stderr'] if item in result and to_text(result[item]) != '')
         for item in items:
             msg = result[item] if msg == '' else '\n'.join([msg, result.pop(item, '')])
 
@@ -79,9 +82,9 @@ def display(obj, result):
 
     # Must pass unicode strings to Display.display() to prevent UnicodeError tracebacks
     if isinstance(msg, list):
-        msg = '\n'.join([to_unicode(x) for x in msg])
+        msg = '\n'.join([to_text(x) for x in msg])
     elif not isinstance(msg, unicode):
-        msg = to_unicode(msg)
+        msg = to_text(msg)
 
     # Wrap text
     msg = '\n'.join([textwrap.fill(line, wrap_width, replace_whitespace=False)
@@ -102,7 +105,7 @@ def display(obj, result):
     else:
         if not first:
             display(hr, 'bright gray')
-        display(msg, 'red' if failed else 'bright purple')
+        display(msg, 'red' if obj.task_failed else 'bright purple')
 
 def display_host(obj, result):
     if 'results' not in result._result: